requirements.txt
pyrogram
tgcrypto
python-dotenv
SQLAlchemy
ormar
aiosqlite